1.0.1.0
Photo Friends Web Application

Features:
Added user profile pages.
Users profile pages can be set to private or public (an option needs to be added to 'settings.php' to set to public or private).
Admins and moderators can view all profiles even if set to private.
Profile pages display the users information, such as, email, title and a small description. As well as, all of their posted images.

See 'ThingsToDo.txt' for a list of goals I want to achieve.